# crates/eidolons-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eidolons API",
    "description": "Privacy-transparent AI proxy API with inline verification metadata",
    "contact": {
      "name": "Mike Marcacci"
    },
    "license": {
      "name": "",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/account": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account — retrieve account info (authenticated).",
        "operationId": "get_account",
        "responses": {
          "200": {
            "description": "Account info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      },
      "post": {
        "tags": [
          "Linked"
        ],
        "summary": "POST /v1/account — create a new account.",
        "operationId": "create_account",
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/account/balances": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account/balances — get credit balance breakdown (authenticated).",
        "operationId": "get_balances",
        "responses": {
          "200": {
            "description": "Balance breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
    "/v1/account/checkout": {
      "post": {
        "tags": [
          "Linked"
        ],
        "summary": "POST /v1/account/checkout — create a checkout session (authenticated).",
        "description": "Accepts any active Stripe price ID. Automatically determines whether to\ncreate a subscription or one-time payment checkout based on the price type.\nFor subscription prices, enforces the one-active-subscription constraint.",
        "operationId": "create_checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Already subscribed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
    "/v1/account/ledger": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account/ledger — get credit ledger entries (authenticated).",
        "operationId": "get_ledger",
        "parameters": [
          {
            "name": "reason",
            "in": "query",
            "description": "Comma-separated reason filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque pagination cursor from a previous response",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max entries to return (default 50, max 200)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ledger entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
    "/v1/account/subscription": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account/subscription — get subscription details (authenticated).",
        "operationId": "get_subscription",
        "responses": {
          "200": {
            "description": "Subscription details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Stripe customer or no subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
...
    "/v1/prices": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "GET /v1/prices — list available prices.",
        "operationId": "list_prices",
        "responses": {
          "200": {
            "description": "List of available prices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPricesResponse"
                }
              }
            }
          },
          "503": {
            "description": "Stripe not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalancePool": {
        "type": "object",
        "required": [
          "amount",
          "source"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string"
          }
        }
      },
      "BalancesResponse": {
        "type": "object",
        "required": [
          "available",
          "pools"
        ],
        "properties": {
          "available": {
            "type": "integer",
            "format": "int64"
          },
          "pools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalancePool"
            }
          }
        }
      },
...
      "CheckoutRequest": {
        "type": "object",
        "required": [
          "price_id"
        ],
        "properties": {
          "cancel_url": {
            "type": "string"
          },
          "price_id": {
            "type": "string"
          },
          "success_url": {
            "type": "string"
          }
        }
      },
      "CheckoutUrlResponse": {
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "properties": {
          "checkout_url": {
            "type": "string"
          }
        }
      },
...
      "CreateAccountResponse": {
        "type": "object",
        "required": [
          "account_id",
          "secret",
          "created_at"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message",
          "type"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "An error response in OpenAI format.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "GetAccountResponse": {
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "stripe_customer_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "LedgerEntry": {
        "type": "object",
        "required": [
          "id",
          "delta",
          "reason",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "delta": {
            "type": "integer",
            "format": "int64"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "token_credits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "token_epoch": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "LedgerResponse": {
        "type": "object",
        "required": [
          "data",
          "has_more"
        ],
        "properties": {
          "cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntry"
            }
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
...
      "ListPricesResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceResponse"
            }
          }
        }
      },
...
      "PriceResponse": {
        "type": "object",
        "required": [
          "id",
          "product_name",
          "currency",
          "type",
          "credits"
        ],
        "properties": {
          "credits": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lookup_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_description": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_name": {
            "type": "string"
          },
          "recurring": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RecurringResponse"
              }
            ]
          },
          "type": {
            "type": "string"
          },
          "unit_amount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "RecurringResponse": {
        "type": "object",
        "required": [
          "interval",
          "interval_count"
        ],
        "properties": {
          "interval": {
            "type": "string"
          },
          "interval_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SubscriptionResponse": {
        "type": "object",
        "required": [
          "id",
          "status",
          "management_url"
        ],
        "properties": {
          "current_period_end": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "management_url": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic",
        "description": "Account credentials. Username is the account UUID, password is the secret returned at account creation."
      }
    }
  },
...
}